Move firebase files to firebase dir
