+++ src/firebase/firestore.js
@@ -1,5 +1,5 @@
-import firebase from '../firebase-config';
-const database = firebase.firestore()
+import firebase from './firebase-config';
+const database = firebase.firestore();
 
 export default (data, uid) => 
   database
